[level 2] Title: 피보나치 수, Time: 3.13 ms, Memory: 78.3 MB -BaekjoonHub

+++ 프로그래머스/2/12945. 피보나치 수/피보나치 수.java
@@ -1,5 +1,5 @@
 class Solution {
-    public long solution(int n) {        
+    public int solution(int n) {        
         int[] fibo = new int[n + 1];
         
         fibo[0] = 0;
